Tweaked it a little to be more general
Modified the script so it is more automated. Meaning, you only have to input necessary stuff like goal and start nodes or map boundary. The rest is taken care of by algorithm.

ALSO. I have changed the positions of the functions so it makes more sense. However, it does not influence functionality whatsoever. That is why there are so many changes in this script. Most of them are insignificant cosmetic changes.